+++ src/pie/bie/graph/index.ts
@@ -8,3 +8,4 @@ export * from "./entityResolver";
 export * from "./edgeProposalQueue";
 export * from "./graphQueryService";
 export * from "./relationshipExtractor";
+export * from "./graphInferenceEngine";